\name{exams2ilias}
\alias{exams2ilias}
\title{Generation of Exams in ILIAS Format}
\description{
  Placeholder manual page for the initial package skeleton.
}
\usage{
exams2ilias(...)
}
\arguments{
  \item{\dots}{Arguments passed to the implementation.}
}
\value{
  An object returned by the eventual ILIAS export implementation.
}
\keyword{utilities}
